Recovering from data type errors

+++ sql/views/07_wide.sql
@@ -8,17 +8,13 @@ SELECT
     file_end_dt,
     layout_id,
     derived_row_serial,
-
     MAX(CASE WHEN canonical_field = 'transaction_date' THEN raw_value END) AS transaction_date,
-
     MAX(CASE WHEN canonical_field = 'description' THEN raw_value END) AS description,
-
     COALESCE(
         MAX(CASE WHEN canonical_field = 'amount' THEN raw_value END),
         MAX(CASE WHEN canonical_field = 'credit_amount' THEN raw_value END),
         MAX(CASE WHEN canonical_field = 'debit_amount' THEN raw_value END)
     ) AS amount,
-
     TRY_CAST(
         NULLIF(
             REGEXP_REPLACE(
@@ -35,15 +31,11 @@ SELECT
                 '-$1'
             ),
             ''
-        ) AS double
+        ) AS DECIMAL(18,2)
     ) AS amount_d,
-
     MAX(CASE WHEN canonical_field = 'check_number' THEN raw_value END) AS check_number,
-
     MAX(CASE WHEN canonical_field = 'status' THEN raw_value END) AS status
-
 FROM canon_long
-
 GROUP BY
     source_file_path,
     source_file_name,
@@ -53,7 +45,6 @@ GROUP BY
     file_end_dt,
     layout_id,
     derived_row_serial
-
 HAVING
     MAX(CASE WHEN canonical_field = 'transaction_date' THEN raw_value END) <> 'Date'
     AND COALESCE(
@@ -61,4 +52,3 @@ HAVING
         MAX(CASE WHEN canonical_field = 'credit_amount' THEN raw_value END),
         MAX(CASE WHEN canonical_field = 'debit_amount' THEN raw_value END)
     ) NOT IN ('Amount', 'Credit', 'Debit');
-
